socia/requirements.txt
flask
flask-session
mysql-connector-python
icecream
python-dotenv
werkzeug
requests